refactor: establish clear boundaries between CLI, harness, and MCP layers
This commit introduces a structural refactor to separate the codebase into distinct boundaries for `dotcontext/cli`, `dotcontext/harness`, and the Model Context Protocol (MCP). Key changes include:

- Updated package.json to define subpath exports for CLI, harness, and MCP.
- Created dedicated entry points for CLI and harness services, ensuring that operator-facing commands and reusable runtime logic are clearly delineated.
- Added new agents and plans to facilitate the transformation towards a harness engineering platform.
- Enhanced documentation to reflect the new architecture and boundaries.

This refactor lays the groundwork for future package splits and improves the overall maintainability of the codebase.
