+++ .context/agents/README.md
@@ -1,6 +1,6 @@
 # Agent Handbook
 
-This directory contains ready-to-customize playbooks for AI agents collaborating on the `@dotcontext/cli` codebase.
+This directory contains ready-to-customize playbooks for AI agents collaborating on the current dotcontext codebase, including the transformation work toward `dotcontext/cli` and `dotcontext/harness`.
 
 ## Available Agents
 
@@ -11,6 +11,15 @@ This directory contains ready-to-customize playbooks for AI agents collaborating
 - [Test Writer](./test-writer.md) -- Write comprehensive unit and integration tests
 - [Documentation Writer](./documentation-writer.md) -- Create clear, comprehensive documentation
 - [Performance Optimizer](./performance-optimizer.md) -- Identify performance bottlenecks
+
+## Harness Transformation Squad
+
+- [Agent Systems Designer](./agent-systems-designer.md) -- Define the target agent topology, harness primitives, and product boundaries
+- [CLI Experience Architect](./cli-experience-architect.md) -- Shape the `dotcontext/cli` experience and keep the CLI thin
+- [Harness Platform Architect](./harness-platform-architect.md) -- Define the runtime and API surface for `dotcontext/harness`
+- [Workflow Orchestration Engineer](./workflow-orchestration-engineer.md) -- Reframe PREVC, plans, agents, and skills as the harness control plane
+- [Harness Quality Auditor](./harness-quality-auditor.md) -- Map guides, sensors, evaluations, and operational quality gates
+- [Migration Release Manager](./migration-release-manager.md) -- Plan packaging, compatibility, rollout, and migration sequencing
 
 ## How To Use These Playbooks
 
@@ -25,5 +34,6 @@ This directory contains ready-to-customize playbooks for AI agents collaborating
 ## Related Resources
 
 - [Documentation Index](../docs/README.md)
+- [Harness Transformation Plan](../plans/dotcontext-harness-engineering-transformation.md)
 - [Agent Knowledge Base](../../AGENTS.md)
 - [Contributor Guidelines](../../CONTRIBUTING.md)
