+++ .context/plans/README.md
@@ -1,11 +1,13 @@
 # Plans Index
 
-This directory contains implementation plans for upcoming changes to `@dotcontext/cli`.
+This directory contains implementation plans for upcoming changes to the dotcontext platform, including the split between `dotcontext/cli` and `dotcontext/harness`.
 
 ## Active Plans
 
 | Plan | Status | Summary |
 | --- | --- | --- |
+| [Dotcontext Harness Engineering Transformation](./dotcontext-harness-engineering-transformation.md) | Proposed | Map the work required to split the product into `dotcontext/cli` and `dotcontext/harness`, with harness engineering as the product center |
+| [Harness CLI MCP Boundaries](./harness-cli-mcp-boundaries.md) | Complete | Separate the internal boundaries so MCP is treated as a transport adapter over the harness |
 | [Simplify Interactive CLI](./simplify-interactive-cli.md) | Proposed | Refactor the interactive menu system in `src/index.ts` to reduce duplication and improve maintainability |
 
 ## Plan Lifecycle
