+++ src/services/harness/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Harness service exports.
+ *
+ * These services hold transport-agnostic harness logic that can be consumed
+ * by MCP, CLI, or future adapters.
+ */
+
+export { HarnessAgentsService, type HarnessAgentsServiceOptions } from './agentsService';
+export { HarnessPlansService, type HarnessPlansServiceOptions } from './plansService';
+export { HarnessContextService, type HarnessContextServiceOptions, type HarnessContextInitResult, type HarnessContextPlanScaffoldResult } from './contextService';
+export { HarnessSkillsService, type HarnessSkillsServiceOptions } from './skillsService';
