+++ src/services/mcp/gateway/plan.ts
@@ -7,14 +7,7 @@
  *           syncPlanMarkdown
  */
 
-import * as path from 'path';
-import { WorkflowService } from '../../workflow';
-import {
-  PHASE_NAMES_EN,
-  createPlanLinker,
-  PrevcStatusManager,
-} from '../../../workflow';
-import { GitService } from '../../../utils/gitService';
+import { HarnessPlansService } from '../../harness';
 
 import type { PlanParams } from './types';
 import type { MCPToolResponse } from './response';
@@ -31,164 +24,65 @@ export async function handlePlan(
   params: PlanParams,
   options: PlanOptions
 ): Promise<MCPToolResponse> {
-  // repoPath is guaranteed to be valid by MCP server initialization
   const repoPath = options.repoPath;
-  // Create statusManager for breadcrumb trail logging in step updates
-  const contextPath = path.join(repoPath, '.context');
-  const statusManager = new PrevcStatusManager(contextPath);
-  const linker = createPlanLinker(repoPath, statusManager, true);
+  const service = new HarnessPlansService({ repoPath });
 
   try {
     switch (params.action) {
       case 'link': {
-        const ref = await linker.linkPlan(params.planSlug!);
-
-        if (!ref) {
-          return createJsonResponse({
-            success: false,
-            error: `Plan not found: ${params.planSlug}`,
-          });
-        }
-
-        const service = new WorkflowService(repoPath);
-        if (await service.hasWorkflow()) {
-          await service.markPlanCreated(params.planSlug!);
-        }
-
-        let canAdvanceToReview = false;
-        if (await service.hasWorkflow()) {
-          const gateResult = await service.checkGates();
-          canAdvanceToReview = gateResult.gates.plan_required.passed;
-        }
-
-        return createJsonResponse({
-          success: true,
-          plan: ref,
-          planCreatedForGates: true,
-          canAdvanceToReview,
-        });
+        return createJsonResponse(await service.link(params.planSlug!));
       }
 
       case 'getLinked': {
-        const plans = await linker.getLinkedPlans();
-        return createJsonResponse({
-          success: true,
-          plans,
-        });
+        return createJsonResponse(await service.getLinked());
       }
 
       case 'getDetails': {
-        const plan = await linker.getLinkedPlan(params.planSlug!);
-
-        if (!plan) {
-          return createJsonResponse({
-            success: false,
-            error: `Plan not found or not linked: ${params.planSlug}`,
-          });
-        }
-
-        return createJsonResponse({
-          success: true,
-          plan: {
-            ...plan,
-            phasesWithPrevc: plan.phases.map(p => ({
-              ...p,
-              prevcPhaseName: PHASE_NAMES_EN[p.prevcPhase],
-            })),
-          },
-        });
+        return createJsonResponse(await service.getDetails(params.planSlug!));
       }
 
       case 'getForPhase': {
-        const plans = await linker.getPlansForPhase(params.phase!);
-
-        return createJsonResponse({
-          success: true,
-          phase: params.phase,
-          phaseName: PHASE_NAMES_EN[params.phase!],
-          plans: plans.map(p => ({
-            slug: p.ref.slug,
-            title: p.ref.title,
-            phasesInThisPrevc: p.phases
-              .filter(ph => ph.prevcPhase === params.phase)
-              .map(ph => ({ id: ph.id, name: ph.name, status: ph.status })),
-            hasPendingWork: linker.hasPendingWorkForPhase(p, params.phase!),
-          })),
-        });
+        return createJsonResponse(await service.getForPhase(params.phase!));
       }
 
       case 'updatePhase': {
-        const success = await linker.updatePlanPhase(params.planSlug!, params.phaseId!, params.status!);
-
-        return createJsonResponse({
-          success,
-          planSlug: params.planSlug,
-          phaseId: params.phaseId,
-          status: params.status,
-        });
+        return createJsonResponse(await service.updatePhase(
+          params.planSlug!,
+          params.phaseId!,
+          params.status!
+        ));
       }
 
       case 'recordDecision': {
-        const decision = await linker.recordDecision(params.planSlug!, {
+        return createJsonResponse(await service.recordDecision({
+          planSlug: params.planSlug!,
           title: params.title!,
           description: params.description!,
           phase: params.phase,
           alternatives: params.alternatives,
-          status: 'accepted',
-        });
-
-        return createJsonResponse({
-          success: true,
-          decision,
-        });
+        }));
       }
 
       case 'updateStep': {
-        const success = await linker.updatePlanStep(
-          params.planSlug!,
-          params.phaseId!,
-          params.stepIndex!,
-          params.status!,
-          { output: params.output, notes: params.notes }
-        );
-
-        return createJsonResponse({
-          success,
-          planSlug: params.planSlug,
-          phaseId: params.phaseId,
-          stepIndex: params.stepIndex,
-          status: params.status,
-        });
+        return createJsonResponse(await service.updateStep({
+          planSlug: params.planSlug!,
+          phaseId: params.phaseId!,
+          stepIndex: params.stepIndex!,
+          status: params.status!,
+          output: params.output,
+          notes: params.notes,
+        }));
       }
 
       case 'getStatus': {
-        const status = await linker.getPlanExecutionStatus(params.planSlug!);
-
-        if (!status) {
-          return createJsonResponse({
-            success: false,
-            error: 'Plan tracking not found. The plan may not have any execution data yet.',
-          });
-        }
-
-        return createJsonResponse({
-          success: true,
-          ...status,
-        });
+        return createJsonResponse(await service.getStatus(params.planSlug!));
       }
 
       case 'syncMarkdown': {
-        const success = await linker.syncPlanMarkdown(params.planSlug!);
-
-        return createJsonResponse({
-          success,
-          planSlug: params.planSlug,
-          message: success ? 'Plan markdown synced successfully' : 'Failed to sync - plan or tracking not found',
-        });
+        return createJsonResponse(await service.syncMarkdown(params.planSlug!));
       }
 
       case 'commitPhase': {
-        // Validate required params
         if (!params.planSlug || !params.phaseId) {
           return createJsonResponse({
             success: false,
@@ -196,91 +90,13 @@ export async function handlePlan(
           });
         }
 
-        // Get the plan and phase
-        const plan = await linker.getLinkedPlan(params.planSlug);
-        if (!plan) {
-          return createJsonResponse({
-            success: false,
-            error: `Plan not found: ${params.planSlug}`,
-          });
-        }
-
-        const phase = plan.phases.find(p => p.id === params.phaseId);
-        if (!phase) {
-          return createJsonResponse({
-            success: false,
-            error: `Phase not found: ${params.phaseId}`,
-          });
-        }
-
-        // Default commit message from phase's commitCheckpoint or generate one
-        const commitMessage = phase.commitCheckpoint ||
-          `chore(plan): complete ${phase.name} for ${params.planSlug}`;
-
-        // Default stage patterns to .context/**
-        const stagePatterns = params.stagePatterns || ['.context/**'];
-
-        // Initialize git service
-        const gitService = new GitService(repoPath);
-
-        // Check if repo is a git repository
-        if (!gitService.isGitRepository()) {
-          return createJsonResponse({
-            success: false,
-            error: 'Not a git repository',
-          });
-        }
-
-        // If dry run, just preview what would be committed
-        if (params.dryRun) {
-          const stagedFiles = gitService.stageFiles(stagePatterns);
-          return createJsonResponse({
-            success: true,
-            dryRun: true,
-            planSlug: params.planSlug,
-            phaseId: params.phaseId,
-            commitMessage,
-            coAuthor: params.coAuthor,
-            filesWouldBeCommitted: stagedFiles,
-          });
-        }
-
-        try {
-          // Stage files matching patterns
-          const stagedFiles = gitService.stageFiles(stagePatterns);
-
-          if (stagedFiles.length === 0) {
-            return createJsonResponse({
-              success: false,
-              error: 'Nothing to commit: no files match the stage patterns or all files are already committed',
-            });
-          }
-
-          // Create the commit
-          const commitResult = gitService.commit(commitMessage, params.coAuthor);
-
-          // Record the commit in plan tracking
-          await linker.recordPhaseCommit(params.planSlug, params.phaseId, {
-            hash: commitResult.hash,
-            shortHash: commitResult.shortHash,
-            committedBy: params.coAuthor,
-          });
-
-          return createJsonResponse({
-            success: true,
-            planSlug: params.planSlug,
-            phaseId: params.phaseId,
-            commit: {
-              hash: commitResult.hash,
-              shortHash: commitResult.shortHash,
-              message: commitMessage,
-              filesCommitted: commitResult.filesCommitted,
-              coAuthor: params.coAuthor,
-            },
-          });
-        } catch (error) {
-          return createErrorResponse(error);
-        }
+        return createJsonResponse(await service.commitPhase({
+          planSlug: params.planSlug,
+          phaseId: params.phaseId,
+          coAuthor: params.coAuthor,
+          stagePatterns: params.stagePatterns,
+          dryRun: params.dryRun,
+        }));
       }
 
       default:
