+++ src/services/shared/index.ts
@@ -106,3 +106,10 @@ export {
   getContextPath,
   getIndexFile,
 } from './contentTypeRegistry';
+
+// Execution helpers
+export {
+  minimalUI,
+  mockTranslate,
+  toolExecutionContext,
+} from './executionContext';
